The beginning of major refactoring of how cross-compiling is done.
This commit removes a little hack in pkgs/top-level/default.nix in that
had the result of making lots of native packages be different just
because crossSystem was specified.

After this commit, you can compare pkgs.zlib for two different
versions of pkgs where one has a crossSystem and the other does not,
and the two are the same.  However, the two gccs of those packages are
different, and I need to figure out why.