Build both stages of the GCC cross-compiler using stdenv.
Previously, we were using the bootstrap tools to build
cross-compilers.  The bootstrap tools are threat to the purity of the
Nix system, so they should only be used for building the stdenv and
nothing else.

Also, this commits set langObjC and langObjCpp to false for
cross-compilers; it doesn't make sense for those attributes to depends
on whether the build system for the cross compiler is Darwin or not.